refac ws connection class 1
