+++ TerminalBuddyAgent/ServerStatusIndicator.swift
@@ -9,17 +9,13 @@
 import SwiftUI
 
 struct ServerStatusIndicator: View {
-    @Binding var isConnected: Bool
-    @Binding var testBool: Bool
+    var isConnected: Bool
     
     var body: some View {
         HStack {
             Circle()
                 .fill(self.isConnected ? Color.green : Color.red)
                 .frame(width: 20, height: 20)
-            Rectangle()
-                .fill(self.testBool ? Color.green : Color.red)
-                .frame(width: 20, height: 20, alignment: .center)
         }
     }
 }
@@ -28,8 +24,8 @@ struct ServerStatusIndicator_Previews: PreviewProvider {
     private let isConnected = true
     static var previews: some View {
         Group {
-            ServerStatusIndicator(isConnected: .constant(true), testBool: .constant(false))
-            ServerStatusIndicator(isConnected: .constant(false), testBool: .constant(true))
+            ServerStatusIndicator(isConnected: true)
+            ServerStatusIndicator(isConnected: false)
         }
     }
 }
